full solution and add types.h

+++ src/alloc.c
@@ -1,5 +1,6 @@
+#include <stdarg.h>
 #include <stddef.h>
-
+#include "types.h"
 #include "kernelpanic.h"
 #include "memu.h"
 
@@ -8,12 +9,12 @@
 
 char* ptr = (char*) ARENA_START;
 
-void* malloc_immortal(size_t size, size_t align) {
+void* malloc_immortal(u32 size, u32 align) {
   if (size == 0) return NULL;
   if (align == 0) align = 1;
 
-  if ((size_t) ptr % align != 0)
-    ptr += align - (size_t) ptr % align;
+  if ((u32) ptr % align != 0)
+    ptr += align - (u32) ptr % align;
 
   if (ptr + size > (char*) ARENA_END)
     kernel_panic("ARENA OVERFLOW, last obj size %d bytes with %d align", size, align);
@@ -23,7 +24,7 @@ void* malloc_immortal(size_t size, size_t align) {
   return result;
 }
 
-void* calloc_immortal(size_t size, size_t align) {
+void* calloc_immortal(u32 size, u32 align) {
   void* result = malloc_immortal(size, align);
   
   if (result == NULL) return NULL;
