+++ src/memu.h
@@ -3,7 +3,9 @@
 
 // Memory Utilities
 
-void* memmove(void* dst, const void* src, size_t n);
-void* memzero(void* dst, size_t n);
+#include "types.h"
+
+void* memmove(void* dst, const void* src, u32 n);
+void* memzero(void* dst, u32 n);
 
 #endif
